backend/scripts/requirements.txt
supabase==1.0.3
python-dotenv==1.0.0
